kernel: fix #GP fault in PIT irq
In certain circumstances, where the only task running is exiting, the
current_task pointer would become invalid once the task was freed. This
caused the PIT irq to try to tick the clock of a already freed task,
deferencing a bad pointer.

The validity of the current_task pointer can be ensured by changing it
to the initial_kernel_task in the case where we attempt to free the
current task.