kernel: implement kernel level stack tracing
The kernel now reads its own executable file (/boot/os_2.o) on startup,
and attempts to read and store the symbol table. With this information,
the kernel can report a stack trace through the serial port if an
unrecoverable state is reached.

Ideally, the kernel stack will also be integrated with the user stack
trace generated when a user process faults, and the stack can be read by
accessing /proc/<pid>/stack.