kernel: implement /proc/<pid>/stack for kernel tasks
Unfortunately, we can't just read the kernel stack to generate the stack
trace, since every kernel stack is mapped to the same place in virtual
memory, and doing a context switch is somewhat unreasonable.

The *hack* solution is to map the stack in using the create_phys_mapping
function, and then transform all the stack frame pointers into this
memory.