kernel: rework blocking to support being interrupted by a signal
Now all of the proc_block_* functions return a value used to indicate
whether or not the blocking was interrupted. This allows kernel code to
allocate resources before blocking, and deallocate them upon return.

This introduces a flag task::should_exit, which is set to true when a
task should exit (sent a terminated signal or maybe a different thread
called execve), which is checked upon system call reentry in order to
not exit kernel system calls while they are running. Now, for instance,
the exit system call calls exit_process(), which will set the current
task's should_exit to true. When the function returns, exit will notice
this state change and then yield and set its sched_state to EXITING,
telling the scheduler to clean up.