build: add setup.sh script to setup the project using cmake
There needs to be a complicated wrapper because we need to different
cmake directories, one for each toolchain we build with (one is built
natively to do code generation and to create the initrd).

The script is configurable through following environment variables.
GENERATOR: generator for cmake to use (defaults to ninja if it exists,
           else does make)
ARCH: architecture to build the os for (defaults to x86_64, the only
      currently supported architecture)
OS_2_NATIVE_DIR: directory of native cmake project (defaults to ./native)
OS_2_BUILD_DIR: directory of os cmake project (defaults to ./build)
OS_AS: assembler to pass cmake (defaults to $ARCH-os_2-as)
OS_CC: c compiler to pass cmake (defaults to $ARCH-os_2-gcc)
OS_CXX: c++ compiler to pass cmake (defaults to $ARCH-os_2-g++)

Note that the script has to pass -DCMAKE_C_COMPILER_WORKS=TRUE in order for it to
successfully build the first time, since the tests rely on linking
against libc which happens to not be built yet, so things may go very
wrong if incorrect AS/CC/CXX are passed.