build: build the iso file by default
In order to correctly track the dependencies of the iso file (it depends
on *everything*), the CMake files now keep track of every buildable
target that gets added, so that dependencies are tracked correctly. This
means that things will _just_ work on matter what target you specify.
This patch also folds the installation step into the main build, so now
building the os requires no additional steps.