kernel: make vm_object reference counting atomic
Out of laziness, the code just took the vm_object lock when modifying
the reference counts. This actually led to a nasty race conditon which
would crash the kernel randomly. How is taking a lock creating a race?

inode_vm_objects take their lock before reading from disk. Reading from
disk is interrupt driven, and so will always sleep with the lock held.
This means that any attempt to aquire this lock can fail, and if it
fails, the code has no choice but to enable interrupts and allow the
currently exucting task to be interrupted.

This is actually a hueg problem, because some kernel functions must
absolutely never be preempted. The prime example is drop_process(),
which will drops a reference to a process and then frees it if needed,
which it usually does since there's only going to be one task. To clean
up the process, this code has to call drop_vm_object on all of the
process's vm_objects. Which calls tries to take the lock, but fails.
In this scenario, the current_task variable points to a complete garbage
half destroyed structure will breaks everything.

This solution to this problem is obvious: use atomic reference counts,
since there's no need to lock (even if the vm_object needs to be freed,
there's no references to it so it doesn't need to be locked. This also
fixes the extremely awkward pattern of whether or not to unlock a lock
after the underlying object is freed).