kernel: fix major procfs issue that prevented ls-ing root, sometimes
The proc file system uses the lower bits of its function pointers to
store additional information about the inode. This is completely
unecessary, but is more space efficent. The issue was that function
alignment was not actually specified on these functions, effectively,
a recipe for disaster. Unforunately, this made the code break
completely, while destroying the backtrace as the stack frame
was corrupted, again, only for certain builds. Since the layout
of the function depends on all the other code, simply adding debug
statements would prevent the issue from occuring.

Luckily, this can all be fixed by just specifying the function alignment
directly through __attribute__((optmize("align-functions=n")))