kernel: keep track of a process's "main thread id"
Lots of things in the system want to default to the main thread when
displaying information (things like /proc/<pid>/stack, for example).
In order to support this, a process has a notion of a main thread
(thread group leader). This simply changed if that task exits, and is
freshly set on fork() and execve().