kernel: fix handling of sigaction(SIGCHLD) with SIG_DFL handler
For some reason I thought that explicitly setting the signal
disposition of SIGCHLD to SIG_DFL should active the SA_NOCLDWAIT.
This not true at all (but might be the case when setting SIG_IGN).

This caused waitpid() to never return when gcc called it after setting
this signal disposition.