libc: fix realloc
After discovering a stupid bug in calloc 7 months ago. I have finally
uncovered this stupid bug in realloc. Of course the only program that
triggered this was gcc...

realloc used to always copy "sz" amount of bytes from the old pointer.
Since it is only reading, there's was never a problem - unless ofcourse
the old pointer was allocated at the end of the heap, in which case we
read past the end of memory and seg fault. This is easily fixed by copying
min(old_size, new_size) amount of bytes.