libc: fix funny printf bug
printf("%d", -1) used to print -00000001 instead of -1. This was
due to implicitly casting the number to an unsigned int. Printf
then though the length of -1 was actually the length of UINT_MAX,
when it only needed to print 1 digit (and the negative).

This was causing ported gcc to fail in the assembly stage, since
gcc output assembly with an offset of '-00000016', which the
assembler didn't like very much.

Considering the amount of warning gcc gives me for comparing ints
with different signs, they really should warn about this type of
thing.