kernel: fix ext2_sync_super_block
This is a remnant from the change that allowed for larger block sizes
in the ext2fs. It always assumed the size of the super block was only
1 block, but in reality, it can be smaller.

Now ext2fs will always write out exactly 1024 bytes, bypassing the
ext2_write_blocks mechanism, instead using fs_pwrite().

This bug fix allows disk images modified by the os to not become
corrupted, and can now be used on a second boot of the system as
well as mounted on a host machine.