libc: fix buffered fwrite()
As an optmiziation, fwrite() will call writev() to flush the buffer
and do the write for sufficently large writes. But the code doing
this was completely incorrect, largely due to incorrectly taking into
consideration that the return value will be larger than the number of
bytes that are supposed to written. This is fixed with proper attention
to details.

This fix *finally* makes ld not fail when writing out an executable, and
thus finally the gcc port successfully compiles hello.c, although it does
take ~40 seconds to complile the 6 lines of code (without even any include
statements).