kernel: implement COW page mappings
On fork(), non-shared writable region's are marked COW in the parent.
The child recieves a cloned vm_object, which holds references to the
parent vm_object's physical pages. These are no longer "owned" by the
parent, since whoever faults first will get their physical page
replaced in the object itself. There are too distinct ways in which
a vm_object is mapped into multiple process's, and they do not mix.
A COW mapping to a MAP_SHARED file mapping can not exist, since the
changes in the vm_object (which is directly shared) will not
propagate to additional processes.

The state of any given mapping is placed directly in the page table
entry (in one of the freely available bits), and this is checked
before deciding to "fix" the mapping by creating a new one.

Currently, COW mappings only occur on fork(), but this mechanism
could be extended to share MAP_PRIVATE page mappings, and maybe
MAP_PRIVATE mappings could COW MAP_SHARED regions if the issue
mentioned above could somehow be solved.