xterm: build xterm with raw vga support
Since xterm was built to be compiled in two different modes,
the easiest way to make it dynamic is to just compile to separate
executables. Now, you can add graphics=1 after the /boot/boot in
grub.cfg, and rebuilding the iso will switch to graphics mode for
the os, without needing to recompile anything.