kernel: make PS/2 mouse driver read input correctly
Due to some sort of inability to count, the PS/2 mouse driver
always read extra 1 extra bytes, which ensured that its information
was never correct.

This patch also implements support for the MOUSE_DOWN and MOUSE_UP
values being reported in the mouse packet. These flags are set only
trigger on the rising/falling edge of a click, and are not continually
set like they are in hardware.