kernel+window_server: support QEMU VM mouse
The VM mouse is built ontop of VMWare's back door functionality, and
expands the number of back door commands. This patch implements support
for this back door, and the VM mouse.

The VM mouse is a PS/2 mouse hack, in both the virtual
machine and in the PS/2 mouse driver. Even though the OS will get the
coordinates from the VM mouse, it still must read information from the
PS/2 mouse, or it will stop recieving information.

In addition, the way that the VM mouse reports its coordinates is differs
from the PS/2 mouse in that the coordinates are absolute (not reported
in deltas). Since the VM doesn't know the screen resolution, these
coordinates are always scaled to be between 0 and 0xFFFF. This scaling must
be converted by whovever renders the cursor, and so kernel mouse_events now
have a scale_mode flag that tells the window server whether or not it needs
to perform a conversion of the coordinate space. To make this easier, the
mouse driver reports the delta between the last two mouse coordinates,
rather than the absolute coordinates it recieves.