kernel: fix race in ext2_lookup
The code was naively checking the value of a lazily initialized
pointer (to see if it needed to initialize it), without taking a
lock on the object. This led to cases where where a lookup could
complete before all the directory entries of that inode were loaded.
This was particuarly noticable when two processes attempted to load
the same font file at the same time, because the Font code asserts
all opens succeed, but they could fail due to this race.