kernel: use block granularity when reading the ext2 inode table
Before, the entire inode table would be stored into memory and read
in all at once. This created a massive allocation, over a megabyte,
and was also very slow. Reading each part as a single block works
much better to elimate this allocation and start up cost, and should
be similarly performant, assuming the block cache works well (which
it doesn't yet, but this change is still good architecturely).