edit: implement undo/redo with document snapshots
The snapshotting mechanism is not the most efficent, but is still
very useful. The current commands will be changed to use a fancy
undo mechanism, but the snapshot mechanism is still useful when
adding new commands or for commands that are very hard to undo
programmatically, like regex find and replace.

Snapshots also offer the possibily of having some form of auto-save
mechanism, so that not everything would be lossed in the event of
some unfortunate accident.