libs: add a library for lexing C source files
This is lexer uses a rather lazy approach to parse things like
number literals and character constants, and also is unable to
determine whether an identifier is a type or not, which is required
to actually parse a C file. Nevertheless, it is pretty decent at
providing syntax highlighting of a source file.