libclipboard: add support for reading the X11 clipboard
The purpose of this library is to be able to copy and paste in the
edit program, but X11 support is very annoying. Since X11 has a custom
event loop, the easiest thing to do is segregate it into it's own
thread, especially since it has to do negotations for anyone to actually
copy from the current program. Unfortunately, Xlib isn't thread safe and
thus only one thread can access X functions at a time.

I think the most elegant solution is to make all the X calls happen in
the background thread and make the normal connection functions interrupt
the event loop with pthread_kill()