kernel: unpolished version of loading devices from the file system
Device files are bound to a specific device with fs_bind_device_to_inode,
which then intercepts future calls to the vfs and directs them to the
device.

Not all devices are generated by the ./makeimg.sh script, and I plan
to change the device numbers to something less random.

Also, the ptmx code is now quite horribly factored. The most sensible
fix is to add a pseudo-terminal file system which can interact with
the file system in a non-hackish way.