meta: add group-based permissions to server sockets and some devices
Now that the kernel supports groups, we can take advantage of them
by making servers run as individual users and allow accessing their
sockets/files only to those in their group. For instance, now a user
must be a member of the dns group in order to access the lookup
server. For window server, this restriction also applies to the
shared memory it uses to communicate bitmaps.