parser: support generating value headers
LIIM::Variant is always used in conjunction with the parser in order
to handles different objects. This is the first step towards making
the parsing mechanism compatible with the c++ type system. This patch
generates a header that defines a Variant with all the non terminals
specified by the grammar, and a blanket type that encompasses all
terminals.

The goal is to have the parser call the reduction callbacks with the
correct type, which will greatly simplifiy the code necessary. It may
even be possible to automatically generate this logic as well, since
it could just call user-defined constructors for the given types.