parser: make reduce functions optionally typed
Before, a function like `reduce_clause$actionlist()' would have a
return type of ModeValue (which is a Variant of several types), and
a single parameter of this same type. Now, if a value header is
specified, the parser generator will create code like this:

        virtual Clause reduce_clause$actionlist(Actionlist&) = 0;

This is much better than before, since the dealing with the Variant
type is completely hidden to users of the generic parser.

This patch also makes use of the move semantics added to the Stack
container, which means that the cost of moving around Vectors while
parsing will hopefully be much lower.