parser: generate a result() function if possible
Every grammar has exactly one start rule, which produces a given type.
The parser now generates a function that type casts the underlying
variant at the top of the value stack to the result type defined by
the grammar. This will only work if the parse method has been called
and returned true.