libapp: allow widget's to specify a preferred size
This is used by Layout code to properly size widgets. The size has
both a width and height, which both default to Size::Auto. For
BoxLayout, this means that space in the layout direction will be
split evenly with other auto sized widgets, and space in the cross
direction will be filled completely. An alignment mechanism should
be added for when this cross size is not auto, as right now it will
simply be top/left aligned.

If the total preferred size of all widgets is greater than the size
of the container widget, preferred size is ignored, and space gets
split evenly between all widgets.