kernel+window_server: report absolute mouse coordinates correctly
Before, the vmware backdoor tried to give the mouse coordinates in
terms of a delta x and y, like the hardware (PS/2 mouse) does. This
doesn't actually work that well, since the delta eventually got out
of sync. The fix is just to report these absolute coordinates directly,
which can then be used to draw the mouse in the perfect location at
all times.