kernel: add basic kernel RNG and add /dev/urandom device
The kernel now detects the presence of the rdrand instruction using
cpuid, and will use this instruction if its available. QEMU will
not support this instruction unless hardware virtualization is enabled
(and the host machnie supports rdrand), so it usually won't be enabled.

The fallback algorithm is just the libc rand() function, which is seeded
by the time read from CMOS. This is definitely terrible, but does produce
kind of random data.

Note that /dev/random is a symlink to /dev/urandom.