kernel: use the DHCP protocol to retrieve critical network params
Network interfaces now have a configuration context, which is managed
by the DHCP code, that specifies the initialization state for the
network interface. Each interface must separately go though the
DHCP process. The kernel's client supports a small subset of the
DHCP options, doesn't care about ip lease timeouts, and has no
retry mechanism. Nevertheless, the code is able to successfully
fetch a dynamic ip and the router's information on a working
network.