kernel: support truncating a mmap file in the tmpfs
If the truncate caused the tmpfs to allocate a new buffer, the inode
vm object it has cached is purged (but old mappings still have keep
a reference to this object). New requests to mmap the file will give
an updated inode vm object, while the old object is takes ownership
of the tmpfs allocated buffer (and frees it when it becomes unmapped).

This mechanism will be used in order to implement window resizing.