window_server: add a window resize mechanism
For testing, typing q or w causes the active window to resize. The
server first resizes and remaps the shared memory the window server
client draws into, and then sends the client a message telling the
client to remap their buffers, and update with the larger size.

A current issue with this approach is that the pixels from the old
bitmap are discarded (which causes screen flickering). This can be
fixed by adjusting the new bitmap appropriately.