libgraphics: add in place size conversion helper functions
This patch adds two functions that adjust the pixels of a bitmap
to fit a new size. shrink_width() shortens the buffer's width, and
adjust_for_size_change() expands the width of a buffer (by moving
over pixels), and clears any pixels not in the old buffer. Another
function, clear_after_y() is used when no width adjustment is
necessary.

These functions work in place on the bitmap so that additional
allocations for copying are unecessary.