window_server: implement resizing with the mouse
This code doesn't really work, since it will almost always crash
if a resize is actually attempted (resizing happens way too often
and can sometimes cause the client to do an invalid mmap).

The solution will be to make a resizes synchronized by the client.
The window server will tell the client it has been resized, but the
client will have to ask the window server itself to find out what
the value is. Once the window server recieves this message, it can
safely mess with the shared memory, and can also combine several
resizes into one. The client will block until it recieves a response,
so things cannot ever become out of sync. This additional overhead
is actually desireable since it will mean less work overall and
of course make the code not prone to data races.