window_server+libapp: make window resizing mostly functional
My added the synchronization mechanism describes in the last commit,
window resizing doesn't result in dumb crashes. However, the kernel's
allocator performs terribly with all the ftruncate calls on the tmpfs,
and the client/server mechanism for blocking is very fragile (the
client must recieve a specific message or it gives up, when really
it should but those messages in some sort of queue). Also, the
terminal completely ignores resize requests and the graphical text
editor doesn't function properly after resizes.