kernel: add kernel level vm region allocator
This patch adds a convenient API for allocating multiple pages in
the kernel. Before, code would often call aligned_alloc() to achieve
this effect, but this is very annoying for the current allocator to
deal with. Now, any code that wants page aligned memory can call
this new API, which will be much more efficent than using malloc().