libws: queue messages from the window server before processing
Certain requests clients make of the window server require the client
to block until they get a response. In the mean time, they might get
the wrong type of response. Therefore, the code now uses a queue of
messages in order to allow the wrong message to be recieved in this
waiting period. Furthermore, it means that multiple messages can be
recieved at the same time (before the client just discarded any extra
messages).