kernel: add preliminary support for the local APIC
This patch doesn't make use of the local APIC yet, but adds a data
structure that models the APIC and a way to enable it. Before the
local APIC can be used though, the IO APIC's must be supported and
thus the kernel interrupt mechanism needs to become more generic.