kernel: use a generic interrupt handler for all irqs
The code nows sets up the IDT to have all 255 entries pointing to
a assembly macro generated entry point, which saves the registers
and exception error codes before calling the generic kernel entry
point.

At this point, the kernel looks up the irq entry in a table and then
dispatches the irq correctly. To make the interface more uniform,
CPU exception error codes are removed from the stack and passed as
an argument to the handler.

This is the first step to make irq handling more generic, as now
the kernel has complete control over all interrupts. The next step
is to abstract over the PIC, so that the IO APIC can replace it
seamlessly. This will include generalizing the irq mappings that
currently take place.

An issue with this new setup is that whether or not an irq should take
place on the alternate stack must be known at compile time and
cannot be configured. This would be fine if only things like
double faults needed to take place on a separate stack (they really
do), but currently scheduling takes place on a separate stack as well.
This is due to every task having an overlapping kernel stack, which
is very convenient for some things but really bad when doing process
tear down (since there's no way to unmap the stack the process is
actually using). It may be a good idea to make the kernel stacks
not overlap, although doing so probably adds some complexity to the
gdt and tss.

Another issue this new irq mechanism doesn't solve is splitting the
irqs between different CPU cores, but that is something that must
be solved in the future.