kernel: use the APIC instead of the PIC
This patch implements support for IO APICs, specified according to
the MADT ACPI table. The IO APIC maps its IRQs according to the
any specified interrupt source override fields in that header, and
assumes all other IRQs are mapped 1 to 1. This is enough to get
the IO APIC to work as a replacement for the PIC, but there are
still some issues in the code.