kernel: fix PIT ticking twice as often
For whatever reason, the PIT sends interrupts twice as often when
the APIC is used. This suggests that the APIC is somehow configured
incorrectly, but there isn't any way to make interrupts happen on
both the rising and falling edges of the interrupt anyway, so I'm
unsure how it is even misconfigured (other interrutps don't happen
twice, just the PIT).

This patch makes the PIT interval 2ms instead fo 1ms if we are using
the APIC (which results in an interval of 1ms always). Eventually the
PIT should become unused as the APIC or HPET timers are used instead.