kernel: make process protected by a mutex instead of a spinlock
Unfortunately, the patch did not change very much (meaning most code
is lazy and doesn't take the process lock before modifying it), but
using a mutex is much better than using a spinlock. A separate spinlock
is needed for both user level mutexes, and the main tid (which is an
artifact of doing process tear down in the scheduler, where the main
tid may be updated but the code cannot sleep).