kernel: limit only the scheduler to use the global list of tasks
The process struct now has its own linked list of tasks, which means
that the scheduler's list of tasks is no longer needed for sending
signals or task lookup. This is in preparation for having multiple
schedulers (one per CPU), so that there will be no global list of
all tasks.

There are still many issues that can affect multi processor scheduling
that this patch doesn't fix. The issue primarily comes down the code
modifying a different task's task_state. When only a single processor
is running, that other task cannot be concurrently executing and thus
the scheduler is going to pick up any changes. But if that task is
currently being scheduled by on some other processor, any writes to
the task state could be clobbered.