kernel: move task and process teardown into a separate process
This works by having the task scheduler notice any tasks that want
to exit and then the scheduler puts them into a queue of tasks to
be destructed. Hopefully giving the responsibility to the scheduler
will make things easy to coordinate.

This patch allows process teardown to block, which should make things
a lot more efficent.