kernel: remove the need for 2 stage bootstrapping
The code used to load the kernel as a multiboot2 module, which a
dedicated boot program prepared and loaded. Much of this preparation
was enabling long mode and setting up a higher half kernel.

This patch moves this logic into the main kernel executable, and
completely removes this old bootstrap program. The reason things
were separated to begin with was that mixing the higher half kernel
addresses with the long mode bootstrapping is not possible. The linker
thinks all symbols are mapped at a very high 64 bit address and thus
refuses to link. The solution is to put all of the bootstrap code and
data in the same section (so that relative jumps are possible), and to
manually adjust every single addresses used in the program. The in
effective achieves linking by hand, given a hard coded starting address
for the executable (which is passed the multiboot2 boot loader).