kernel: fix generating a default floating point state
This patch adds an exception handler for the #XF, which just triggers
an SIGFPE signal. When in kvm mode, most floating point exceptions
trigger a precision exception, which was not properly masked. The issue
was that the code that generated a default FPU state using fninit() called
fxrstor() instead of fxsave(), meaning the result of fninit() was never
used. This caused kvm to send simd exceptions, which are now masked.