kernel: fix loading of the gdt and idt when initializing APs
The issue seems to have been that the old code did not derefernce the
pointers to the gdtr and idtr, which caused them to be invalid. Qemu
without kvm seemed to ignore this, but qemu with kvm got very upset
and would effectively hang the emulator. With this change, the os
seems to actually work just fine with kvm enabled.