kernel: slightly tweak the processor struct
This patch now assigns processor ids sequentially, instead of going
off the acpi tables. This means the numbers should be next to each
other and be suitable for array access. This patch also begins moving
fields that used to be global into the processor struct, like the
fallback kernel stack used for handling double faults.