kernel: store the current processor in GS base MSR
Just like how the current userspace thread is stored in the FS base,
the current processor should be stored in the GS base. The kernel
should use the swapgs instruction to hide this address from userspace,
but that can be a little complex, so this field is visible to userspace
for now.