kernel: fix get_current_processor() implemenation
This code forgot to clobber the rax register. For any normal
function, this is irrelevant, as the return value is in rax so the
compiler "automatically" clobbers it. However, since the code was
inlined, there was no function call and thus the compiler didn't
save the registers, resulting in horribly broken code.