kernel: refactor scheduler data structures into the processor struct
Instead of static fields for the scheduling list and the current task,
this state is now stored in the processor struct, meaning that each
processor gets its own current task and scheduler. Now all that's left
is to support SMP is to actually put tasks in different processors
run queues.