kernel: support SMP
This patch changes the old sched_add_task and sched_remove_task to
be prefixed with the work local, and makes sched_add_task distribute
the tasks between the different CPUs.

Every task is pinned to the CPU it starts on, and its CPU is governed
by an atomically increasing counter modulo the number of CPUs.

Tasks get scheduled by sending an IPI to the processor (if necessary),
where the processor will then add the task to the scheduler's queue
with interrupts disabled. This allows the scheduler to not have to
take locks when scheduling, but this method could actual be less
efficent due to the overhead of sending an IPI (but this way is less
prone to race conditions).

In the current implemenation, the operating system fails to boot around
1/3 of the time with 2 CPUs, and will always fail if there are more than
2 CPUs. This is due to flaws in the CPU message passing mechanism.
Additionally, signal dispatch is liable to crash the kernel (since it
doesn't use the locks it should), but at least threading capabilites
work as well as everything else.

A lot of the procfs information now needs to be updated to have per-CPU
information for both tasks and overall scheduling statistics. Likewise,
programs like top need to support more than 1 CPU.