kernel+userland: properly support termios VERASE
Before I though making the backspace's ASCII 127 was the problem
with vim, but in actuality, it was that VERASE was improperly set
to '#' (which meant that # was interpreted as a backspace character).

Since VERASE can be any number of things (it is 127 on linux, but can
also be 8), portable code should use its value when determining
whether or not a key is the backspace. This makes edit work correctly
on linux.

Since ASCII BS makes more sense for the backspace key than ASCII DEL,
there's no reason to change VERASE and the backspace keycode back to
127, as the only thing that matters is that the two are same.