loader+libdl: make dynamic objects ref counted
This patch makes all dynamic objects reference counted, and makes
objects drop all their references to their dependencies. This makes
the cleanup handling in dlclose() a lot better.

Unfortunately, this solution does not work properly, as it does not
account for relocations. For RTLD_GLOBAL objects, non-dependencies
can hold references to the object, and would not work properly in
cases where that object is dlclosed. This is rare enough that it
can be ignored for now, though.