kernel+libc: use syscall/sysret instructions for system calls
This is supposed to be faster than directly doing an irq, although
user code could still do that if it wants. Since syscall doesn't save
the user stack, it gets saved in an used field of the per processor
tss (this makes sense since the tss has to be read anyway to locate
the kernel stack).