kernel: fix profile sampling
The actual context we want to use (when pre-empted), is the irq
state saved on the stack. This gets saved after the sample is taken,
so it has to be used rather than the stored value. In cases where
there is no such irq stack, the kernel fakes one by getting the
current rip and rbp.