libc: add a generic mechanism for parsing files
This parser will be used for all of the netdb get,set,end functions,
and hopefully could also be used to parsing /etc/passwd and /etc/group.

This patch makes the parser for /etc/hosts properly handle empty lines
and comments, which will also be allowed for anything else that uses
this.