kernel: fix broken clock/timer code
Clocks are now locked before incremented, and as such don't have
race conditions when running with SMP. Additionally, the ticking
of timers accounts for the fact that they can be removed during
iteration.

In addition, the process's alarm timer is explicitly maintained
accross execve(), which is also tested by alarm_test.