kernel: implement and use a simple LRU block cache
This dramatically increases the performance of large executables.
For example, gcc takes ~15 seconds to compile hello.c uncached, and
only ~1 second when cached. This clearly implies that the uncached
path needs to be optimized, but gcc actually becomes useable cached.

Additionally, the kernel memory management needs to take advantage of
the cache's existance, and lessen the size of the cache whenever an
OOM condition is detected.