kernel: add kernel level timer facility
Any code in the kernel can now register a callback to be called after
a timeout. This timer can be cancelled or changed later. If the timer
is only supposed to be called once, the callback is responsible for
freeing the timer, or otherwise reseting its value so it can be called
again.