kernel: implement getrusage()
This patch also fixed the lack of child resource count inheritance
caused by the rewriting of the waitpid() mechanism. This menas that
the times() function now properly keeps track of children, as well
as getrusage(RUSAGE_CHILDREN).

Currently, only the time fields of struct rusage are being kept track
of, but most of the other fields are easy to implement.