ports: add (broken) port of cmake
Because cmake is a build system that builds itself, porting it is
very difficult. In order to make cmake think os_2 is a real operating
system, a new host version of cmake must be built first. Then that
cmake can be used to build the ported cmake.

In addition, cmake makes use of libstdc++, which fails to provide
basic functions since libstdc++ doesn't like libc.

Cmake may have been broken only because no there was no gcc on the
system, but there most likely are a ton of other issues that may
prevent cmake from working properly. It was a challenge just to get
cmake the compile.