kernel: automatically shrink the block cache when needed
This patch fixes the kernel's accounting of the total available
memory (but not the per-process accounting). The kernel then is
able to use this accurate accounting of physical memory to attempt
to reclaim memory from the block cache.

The main issue with this mechanism is that there is that the block
cache will continually try to cache pages even under memory pressure.
Ideally, it should just stop caching once a certain amount of memory
is used.

Additionally, the kernel needs an OOM killer (and potentially a disk
swap mechanism) to deal with the cases where trimming the block cache
is insufficent.