window_server: make windows have a tree hiearchy
When implementing things like context menus and modal dialogues,
client code makes a child window. This child window has the same
order relative to other window trees as the parent, but is always
drawn in front of the parent window. In addition, clicking the child
window brings the parent into view, and clicking the parent brings
the child into view.